Tomato: support of corners sliding for mouse sliding operator
This commits re-enables mouse sliding operator which behaves in the same
way as it used to for search area and marker position, but for pattern
it's now using individual corners. This allows to define affine transformation
which later would be used by tracker algorithm.

For non-affine tracking code is also ported but currently it's hardcoded to
use affine transformation. When real affine trackers and RNA/DNA changes would
be commited, easy to switch sliding operator to old-school behavior for
non-affine patterns.