RNA C++: collections functions are now properly handled, meaning it is now possible to call BlendData.meshes.remove() directly.
It requires special order of classes in header file (namely classes used for
collection functions should be declared before collection properties with this
classes are declared). Currently used naive linear search for this, could be
replaced with hash or bisect, but for now performance here is not an issue.

Patch by Sergey, but committing now because it's needed for next commit.