Fix for theme colors not correctly applied to the user interface.
This commit partly reverts the changes in revision 53197 for
making Freestyle optional.  Specifically #ifdef WITH_FREESTYLE
was removed from source/blender/editors/include/UI_resources.h so
that Freestyle-related theme properties are always defined
regardless of whether Freestyle is enabled or not.  This is in
line with the fact that Freestyle-related DNA components are
always defined no matter how a build is configured.